Box3D v1.0
When I was in NC computing we learned Visual Basic (4.5+).
Our teacher told us that we were going to be learning how to create a 3D box and making it spin on the x axis.
So, between the Tuesday and the Thursday I created a faux 3D box effect.
This is a re-implementation of that ancient code in shiny new JS.